Backend/requirements.txt
Django==4.2.27
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
django-cors-headers==4.3.1
django-filter==24.2
dj-database-url==2.1.0
psycopg2-binary==2.9.9
gunicorn==22.0.0
whitenoise==6.7.0
Pillow==10.3.0
